requirements.txt
flask>=3.0
python-dotenv>=1.0
pyyaml>=6.0
chromadb>=1.5
